requirements/requirements-py38-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --extra dev --emit-index-url --python-platform linux pyproject.toml --output-file=requirements/requirements-py38-dev.txt
--index-url https://pypi.org/simple
--extra-index-url https://artifactory.spotify.net/artifactory/api/pypi/pypi/simple/

absl-py==2.1.0
    # via
    #   tensorboard
    #   tensorflow
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.10
    # via gcsfs
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
astunparse==1.6.3
    # via tensorflow
async-timeout==4.0.3
    # via aiohttp
attrs==24.2.0
    # via aiohttp
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via typer
coverage==7.6.4
    # via pytest-cov
decorator==5.1.1
    # via gcsfs
dnspython==2.7.0
    # via coglib (pyproject.toml)
docstring-parser==0.16
    # via google-cloud-aiplatform
exceptiongroup==1.2.2
    # via pytest
fastavro==1.9.7
    # via coglib (pyproject.toml)
filelock==3.16.1
    # via
    #   torch
    #   triton
flatbuffers==24.3.25
    # via tensorflow
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.10.0
    # via
    #   gcsfs
    #   torch
gast==0.6.0
    # via tensorflow
gcsfs==2024.10.0
    # via coglib (pyproject.toml)
google-api-core==2.22.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.36.0
    # via
    #   gcsfs
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   tensorboard
google-auth-oauthlib==1.2.1
    # via
    #   gcsfs
    #   tensorboard
google-cloud-aiplatform==1.71.1
    # via coglib (pyproject.toml)
google-cloud-bigquery==3.26.0
    # via
    #   coglib (pyproject.toml)
    #   google-cloud-aiplatform
google-cloud-bigquery-storage==2.27.0
    # via coglib (pyproject.toml)
google-cloud-core==2.4.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-resource-manager==1.13.0
    # via google-cloud-aiplatform
google-cloud-storage==2.18.2
    # via
    #   coglib (pyproject.toml)
    #   gcsfs
    #   google-cloud-aiplatform
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-pasta==0.2.0
    # via tensorflow
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.65.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.13.1
    # via google-cloud-resource-manager
grpcio==1.67.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   tensorboard
    #   tensorflow
grpcio-status==1.62.3
    # via google-api-core
h5py==3.12.1
    # via tensorflow
httplib2==0.22.0
    # via oauth2client
idna==3.10
    # via
    #   requests
    #   yarl
importlib-metadata==8.5.0
    # via markdown
importlib-resources==6.4.5
    # via coglib (pyproject.toml)
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via torch
keras==2.15.0
    # via tensorflow
libclang==18.1.1
    # via tensorflow
markdown==3.7
    # via tensorboard
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.3.2
    # via tensorflow
more-itertools==10.5.0
    # via coglib (pyproject.toml)
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy==1.13.0
    # via coglib (pyproject.toml)
mypy-extensions==1.0.0
    # via mypy
networkx==3.2.1
    # via torch
numpy==1.26.4
    # via
    #   h5py
    #   hnswlib
    #   ml-dtypes
    #   shapely
    #   tensorboard
    #   tensorflow
    #   torchtext
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==8.9.2.26
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.20.5
    # via torch
nvidia-nvjitlink-cu12==12.6.77
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
oauth2client==4.1.3
    # via coglib (pyproject.toml)
oauthlib==3.2.2
    # via requests-oauthlib
opt-einsum==3.4.0
    # via tensorflow
packaging==24.2
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   pytest
    #   tensorflow
pluggy==1.5.0
    # via pytest
propcache==0.2.0
    # via yarl
proto-plus==1.25.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery-storage
    #   google-cloud-resource-manager
protobuf==4.25.5
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery-storage
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
    #   tensorboard
    #   tensorflow
psutil==6.1.0
    # via coglib (pyproject.toml)
pyasn1==0.6.1
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via
    #   google-auth
    #   oauth2client
pydantic==2.9.2
    # via google-cloud-aiplatform
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via rich
pyparsing==3.2.0
    # via httplib2
pytest==8.3.3
    # via
    #   coglib (pyproject.toml)
    #   pytest-cov
pytest-cov==6.0.0
    # via coglib (pyproject.toml)
python-dateutil==2.9.0.post0
    # via google-cloud-bigquery
requests==2.32.3
    # via
    #   gcsfs
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   requests-oauthlib
    #   tensorboard
    #   torchtext
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==13.9.4
    # via typer
rsa==4.9
    # via
    #   google-auth
    #   oauth2client
ruff==0.7.3
    # via coglib (pyproject.toml)
sentencepiece==0.2.0
    # via coglib (pyproject.toml)
setuptools==75.3.0
    # via
    #   tensorboard
    #   tensorflow
shapely==2.0.6
    # via google-cloud-aiplatform
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   astunparse
    #   google-pasta
    #   oauth2client
    #   python-dateutil
    #   tensorboard
    #   tensorflow
sqlitedict==2.1.0
    # via coglib (pyproject.toml)
sympy==1.13.3
    # via torch
tensorboard==2.15.2
    # via tensorflow
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.15.1
    # via coglib (pyproject.toml)
tensorflow-estimator==2.15.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.37.1
    # via tensorflow
termcolor==2.5.0
    # via tensorflow
tomli==2.0.2
    # via
    #   coverage
    #   mypy
    #   pytest
torch==2.3.0
    # via
    #   coglib (pyproject.toml)
    #   torchtext
torchtext==0.18.0
    # via coglib (pyproject.toml)
tqdm==4.67.0
    # via torchtext
triton==2.3.0
    # via torch
typer==0.13.0
    # via coglib (pyproject.toml)
typing-extensions==4.12.2
    # via
    #   multidict
    #   mypy
    #   pydantic
    #   pydantic-core
    #   rich
    #   tensorflow
    #   torch
    #   typer
urllib3==2.2.3
    # via requests
werkzeug==3.1.3
    # via tensorboard
wheel==0.45.0
    # via astunparse
wrapt==1.14.1
    # via tensorflow
yarl==1.17.1
    # via aiohttp
zipp==3.20.2
    # via
    #   importlib-metadata
    #   importlib-resources
